WIP simplifying list component structure and submitted card layout
