+++ client/src/components/desk/RequestedHardwareItem.tsx
@@ -1,5 +1,5 @@
 import React, {Component} from "react";
-import {Button, Card, Header, Icon, Label} from "semantic-ui-react";
+import {Button, Card, Header, Icon, Label, Progress} from "semantic-ui-react";
 import {RequestedItem} from "../inventory/HardwareItem";
 import TimeAgo from "react-timeago";
 
@@ -8,6 +8,7 @@ interface RequestedHardwareItemProps {
     handleApprove: (id: number) => void;
     handleDecline: (id: number) => void;
 }
+
 class RequestedHardwareItem extends Component<RequestedHardwareItemProps, {}> {
     constructor(props: RequestedHardwareItemProps) {
         super(props);
@@ -15,22 +16,22 @@ class RequestedHardwareItem extends Component<RequestedHardwareItemProps, {}> {
 
     public onClickApprove = () => {
         this.props.handleApprove(this.props.item.id);
-    }
+    };
 
     public onClickDecline = () => {
         this.props.handleDecline(this.props.item.id);
-    }
+    };
 
     public render() {
         return (
             <Card>
-              <Card.Content>
-                  <Header size="small" className="hardware-item-user-name">
-                      <Label circular color="blue">
-                          {this.props.item.qtyRequested}x
-                      </Label> {this.props.item.name}
-                  </Header>
-              </Card.Content>
+                <Card.Content>
+                    <Header size="small" className="hardware-item-user-name">
+                        <Label circular color="blue">
+                            {this.props.item.qtyRequested}x
+                        </Label> {this.props.item.name}
+                    </Header>
+                </Card.Content>
                 <Card.Content>
                     <Icon name="wrench"/> 16 in stock, <span style={{color: "red"}}>2 fulfilled</span>
                 </Card.Content>
@@ -40,16 +41,17 @@ class RequestedHardwareItem extends Component<RequestedHardwareItemProps, {}> {
                 <Card.Content>
                     <Icon name="clock outline"/> <TimeAgo date={"August 4, 2019"}/>
                 </Card.Content>
-              <Card.Content extra>
-                <div className="ui two buttons">
-                    <Button color="red" onClick={this.onClickDecline}>
-                        Decline
-                    </Button>
-                    <Button color="green" onClick={this.onClickApprove}>
-                        Approve
-                    </Button>
-                </div>
-              </Card.Content>
+                <Card.Content extra>
+                    <div className="ui two buttons">
+                        <Button color="red" onClick={this.onClickDecline}>
+                            Decline
+                        </Button>
+                        <Button color="green" onClick={this.onClickApprove}>
+                            Approve
+                        </Button>
+                    </div>
+                </Card.Content>
+                <Progress attached="top" percent={5} active={true} color={"green"}/>
             </Card>
         );
     }
