+++ client/src/components/desk/types/Desk.ts
@@ -1,0 +1,3 @@
+export interface Card {
+
+}
