+++ client/src/index.scss
@@ -1,6 +1,6 @@
 @import "~semantic-ui-css/semantic.min.css";
 @import "styles/hardware.scss";
-@import "styles/RequestManagement.scss";
+@import "styles/Desk";
 
 
 body {
