+++ client/src/styles/hardware.scss
@@ -7,7 +7,6 @@
   > .hw-card {
     width: 100%;
     max-width: 500px;
-    margin: 5px;
   }
 }
